+++ sia/belief/gmr.h
@@ -1,9 +1,8 @@
-/// Copyright (c) 2018-2021, Parker Owan.  All rights reserved.
+/// Copyright (c) 2018-2022, Parker Owan.  All rights reserved.
 /// Licensed under BSD-3 Clause, https://opensource.org/licenses/BSD-3-Clause
 
 #pragma once
 
-#include "sia/belief/distribution.h"
 #include "sia/belief/gmm.h"
 
 #include <Eigen/Dense>
@@ -17,7 +16,7 @@ namespace sia {
 ///
 /// References:
 /// [1] http://www.stat.rice.edu/~hgsung/thesis.pdf
-class GMR : public GMM, Regression {
+class GMR : public Inference {
  public:
   /// Initialize GMR from a GMM.  The input indices define which part of the
   /// GMM state vector is an input $x$, and output indices define which slice of
@@ -35,7 +34,7 @@ class GMR : public GMM, Regression {
                std::vector<std::size_t> output_indices,
                double regularization = GMM::DEFAULT_REGULARIZATION);
 
-  /// Performs the regression $p(y|x)$.
+  /// Performs the inference $p(y|x)$
   const Gaussian& predict(const Eigen::VectorXd& x) override;
   std::size_t inputDimension() const override;
   std::size_t outputDimension() const override;
@@ -58,6 +57,7 @@ class GMR : public GMM, Regression {
     Gaussian m_gx;                            // need for local weight
   };
 
+  GMM m_gmm;
   Gaussian m_belief;
   std::vector<std::size_t> m_input_indices;
   std::vector<std::size_t> m_output_indices;